Move: IdeaByStatusChart from components to charts
